Add safeguards against duplicate paths on the rtp
There was a bug where if you tried to maktaba#plugin#Install a plugin that was
already on the runtimepath as a RELATIVE path, then maktaba would also add the
ABSOLUTE version of the plugin path to the runtimepath. This would allow the
plugin to be sourced twice.

Fixing this involved adding a maktaba#rtp#Remove function.