Don't throw errors in async's sync fallback.
If a system call fails under Syscall.CallAsync when async is available,
the exit code, stdout, and stderr are just passed to the handler function,
but if async isn't available, the sync fallback throws a ShellError.

The fallback should forward results to the handler function so you only have to
handle errors in one place.

Closes google/vim-maktaba#121