Add support for Syscall.CallAsync callback that doesn't accept env_dict (#179)
Try a 1-arg callback first and fall back to a 2-arg call on "Not enough arguments" error.
There was a lot of conceptual overhead to the env_dict thing, and it's simple enough to bind state to the callback with `maktaba#function#Create`. Provided an example of how to do that in the Syscall.CallAsync help.